Add offline mode, summary page, notifications
Add offline mode: grades and homework loading is now really fast.
Add the summray page (updated), you can now access your last grades and homework on the main page.
The notifications allow the the user to be informed at the slightest addition of notes. This feature use the background_fetch package with a headless task when the app is closed.